+++ server/src/models/notification.py
@@ -23,6 +23,7 @@ class Notification(Base):
     payload = Column(JSON, nullable=False)
     state = Column(String, nullable=False, default=NotificationState.PENDING.value)
     originator_event_id = Column(Integer, ForeignKey("events.id"), nullable=True)
+    traceback = Column(String, nullable=True)
 
     recipient_member = relationship("Member")
     action = relationship("Action")