+++ server/src/notifications/notifications_consumer.py
@@ -2,6 +2,7 @@ import datetime
 import logging
 import os
 import time
+import traceback
 from typing import Optional
 
 from src import Notification
@@ -15,11 +16,12 @@ class NotificationsConsumer:
         self,
         interval_seconds: int = 10,
         backoff_after_num_attempts: int = 3,
+        session: SessionLocal = SessionLocal(),
     ):
         self._interval_seconds = interval_seconds
         self._backoff_after_num_attempts = backoff_after_num_attempts
         self._logger = logging.getLogger(__name__)
-        self._session = SessionLocal()
+        self._session = session
 
     def run(self):
         """
@@ -65,6 +67,7 @@ class NotificationsConsumer:
         except Exception as e:
             self._logger.exception(f"Failed to send notification: {e}")
             notification.state = NotificationState.FAILED.value
+            notification.traceback = traceback.format_exc()
         finally:
             self._session.add(notification)
             self._session.commit()